+++ packages/core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 0.11.6
+
+### Patch Changes
+
+- 314b396: Builder: improve accuracy of invalid binding detection
+- f442a8c: fix: support for builder custom component with a colon in their name
+
 ## 0.11.5
 
 ### Patch Changes